+++ src/StorageSync/StorageSync/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Migrated Azure AD features in Az.StorageSync to MSGraph APIs. The cmdlets will call MSGraph API according to input parameters: New-AzStorageSyncCloudEndpoint
 * Changed default parameter set of Invoke-AzStorageSyncChangeDetection to use full share detection
 
 ## Version 1.6.1